Move the creategeojsonbloodhound value to the search module
